[LOOKML-4] LookML parser
Add a parser for well-formed LookML.

By "well-formed LookML" we mean syntactically valid LookML
strings. No schema is enforced; so, for example, the root
property does not need to be named "model", or "view"; it
can just as well be called "foo".

The parser is implemented in JavaCC (a Java-based parser
generator that generates top-down parsers). It is a push
parser (like SAX in the XML world) that does not build an
AST but just sends events to an ObjectHandler.

Fixes #4
Close #8