[LOOKML-5] Define schema for LookML
Define a schema that determines whether a LookML string is
valid. The schema is represented by a new
`interface LookmlSchema`, a utility `class LookmlSchemas`,
a builder, etc.

Another way to define a schema is using a string (or file)
in the Schema-LookML dialect. We provide an example in
"mini-lookml-schema.lkml".

Schema-LookML is self-defining. There is a schema file,
"schema-schema.lkml", and also an object,
`SchemaLookml.schema()`.

In addition to defining schema, we provide Mini-LookML, an
example dialect of LookML. New `class MiniLookml` contains
methods that link to the schema file
("mini-lookml-schema.lkml"), the schema object, and an
example model that complies with the schema
("mini-lookml-example-model.lkml").

The example model in fact contains at least one instance of
every property. This is very useful for testing hand-written
code around the model (validators, builders) and we intend
to make it a best practice for anyone writing a schema.

Fixes #5
Close #9