[LOOKML-6] Validate LookML against a schema, and build a typed AST
Provide a schema validator that validates LookML against a
schema (LookmlSchema).

The validator implements ObjectHandler, and can therefore
receive a stream of document events; it sends output to two
new interfaces: PropertyHandler is similar to ObjectHandler
but the events are tagged with their valid property, and
ErrorHandler is for events that do not match the schema.

This change also includes an example of a builder that
creates a typed AST. The `MiniLookml.build()` method will
create an instance of `class MiniLookml.Model`, with nested
instances of `class MiniLookml.View`, etc. We don't recommend
using those classes in production, but if you have written a
schema you can follow that pattern and create your own
builder and AST classes.

Fixes #6
Close #11